make: Add make wrapper for builds
This makes it possible to simply press `make` and have a kernel. It will
download a toolchain, create a configuration, and so forth. The .config
file will only be created if it does not already exist, making it
possible to edit out/.config manually during development. This takes
care of setting all relevant environment variables, and final kernels
will wind up in out/arch/arm64/boot/Image.lz4-dtb. We also automatically
checkout submodules, in case the user forgot to do so.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>